+++ geoelan/src/eaf2geo/mod.rs
@@ -243,6 +243,8 @@ pub fn run(args: &clap::ArgMatches) -> std::io::Result<()> {
         .collect();
     kml_styles.sort_by_key(|e| e.name.to_owned());
 
+    let filename_suffix = format!("_{geoshape_arg}_{}", tier.tier_id);
+
     // Generate KML
     let placemarks: Vec<Placemark> = downsampled_clusters
         .iter()
@@ -255,7 +257,7 @@ pub fn run(args: &clap::ArgMatches) -> std::io::Result<()> {
 
     // Serialize to KML v2.2. No line breaks/indentation.
     let kml_doc = kml_to_string(&kml);
-    let kml_path = files::affix_file_name(&eaf_path, None, Some(geoshape_arg), Some("kml"), None);
+    let kml_path = files::affix_file_name(&eaf_path, None, Some(filename_suffix.as_str()), Some("kml"), None);
 
     match files::writefile(&kml_doc.as_bytes(), &kml_path) {
         Ok(true) => println!("Wrote {}", kml_path.display()),
@@ -268,7 +270,7 @@ pub fn run(args: &clap::ArgMatches) -> std::io::Result<()> {
 
     // Serialize GeoJSON. Not indented (= smaller size for web use).
     let geojson_doc = geojson.to_string();
-    let geojson_path = files::affix_file_name(&eaf_path, None, Some(geoshape_arg), Some("json"), None);
+    let geojson_path = files::affix_file_name(&eaf_path, None, Some(filename_suffix.as_str()), Some("json"), None);
 
     match files::writefile(&geojson_doc.as_bytes(), &geojson_path) {
         Ok(true) => println!("Wrote {}", geojson_path.display()),